use solidus 1.4.2
in order to allow the project to the bundled, we need to use an older version of Solidus, but 2.0 moves to rails 5 so this seems safest for the time being. it also requires deface to be added to the Gemfile.
along with moving to 1.4.2, the gemspec limits solidus to <2.0 (for the same reason as above).
with moving to 1.4.2, some mixins were deprecated and had to be added back in.
we also bump the version of poltergeist use, as 1.5.0 uses a version of phantomJS that is no longer supported